+++ apps/electron-frontend/src/components/my-workflows/import.tsx
@@ -1,6 +1,6 @@
 import { openTabPage } from '@/lib/electron-bridge';
 import { readWorkflowFromFile, readWorkflowFromPng } from '@comflowy/common/comfyui-bridge/export-import';
-import { PersistedWorkflowDocument, documentDatabaseInstance } from '@comflowy/common/local-storage/document-database';
+import { PersistedWorkflowDocument, documentDatabaseInstance } from '@comflowy/common/storage/document-database';
 import { useAppStore } from '@comflowy/common/store';
 import { message } from 'antd';
 import { useRouter } from 'next/router';
